JS: preserve exactness in ++/--, extract arithmetic helpers, route coercion errors through executor.error
- Fix: increment/decrement (i++ / --i) built a fresh JSNumber that dropped the
  exact fraction, so `for (...; i++)` loops reintroduced the order-dependent
  rounding bug (e.g. `i / 7 * 700` gave 100.002). Now routes through
  numberArithmetic so the counter stays exact. Adds regression tests.
- Refactor: move the rational/float arithmetic helpers into a dedicated
  executor/arithmetic.ts; executeBinaryExpression is now just dispatch.
- Cleanup: de-duplicate the PLUS number+number branch (handled identically
  regardless of the coercion flag) and use the guard-clause form in
  arithmeticWithCoercion.
- Errors: raise TypeCoercionNotAllowed via executor.error (translation-driven)
  instead of hand-built message strings; standardises on the both-types format
  and updates the affected test assertions.
- Docs: comment the Fraction zero-denominator invariant guard.
